Implement Livewire component for task management and enhance task display with search functionality
